pyproject.toml
[tool.poetry]
name = "training-spark"
version = "0.1.0"
description = ""
authors = ["john doe <arjan.hulshoff@example.com>"]
readme = "README.md"
packages = [{include = "training_spark"}]

[tool.poetry.dependencies]
python = "^3.10"
apache-airflow-providers-cncf-kubernetes = "^7.4.2"
apache-airflow-providers-apache-spark = "^4.4.0"
pendulum = "^2.1.2"
dbt-postgres = "^1.7.6"
avro = "^1.11.3"
delta-spark = "^3.0.0"
dbt-core = "1.7.6"


[tool.poetry.group.dev.dependencies]
mypy = "^1.7.1"
pytest = "^7.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
